+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changes to Calva.
 
 - [A more flexible evaluate-to-cursor command](https://github.com/BetterThanTomorrow/calva/
 issues/1901)
+- [Test runner does not show stacktrace on error](https://github.com/BetterThanTomorrow/calva/issues/424)
 
 ## [2.0.307] - 2022-10-11